refactor into hook; clean up unused vars
